requirements.txt
asgiref==3.8.1
Django==4.2.13
django-tinymce==4.0.0
Pillow==9.3.0
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
